Modified App component and codition of handleCheck
