+++ src/components/App.jsx
@@ -1,0 +1,26 @@
+import React, { useState } from 'react';
+
+import Page from './Page';
+
+export default function APP () {
+    const [state, setState] = useState({
+        count: 0,
+      });
+    
+      const { count } = state;
+    
+      function handleClick(conditionValue) {
+          setState({
+            count: count + conditionValue
+          });
+          return;
+      }
+    
+      return (
+        <Page
+          count={count}
+          onClick={handleClick}
+        />
+      );
+}
+
