+++ src/components/Buttons.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+
 import Button from './Button';
 
 export default function Buttons({ onClick }) {
