+++ src/components/CounterBtn.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 export default function CounterBtn({ count, onClick }) {
   return (
-    <button type="button" onClick={onClick}>
+    <button type="button" onClick={() => onClick(1)}>
       Click me!
       (
       { count }
