+++ src/components/Page.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+
 import Buttons from './Buttons';
 import CounterBtn from './CounterBtn';
